Move migration and test model back to src

+++ database/migrations/2024_04_21_153207_create_user_table.php
@@ -4,11 +4,16 @@ declare(strict_types=1);
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Schema;
 
 return new class () extends Migration {
     public function up(): void
     {
+        if (App::environment() !== 'testing') {
+            return;
+        }
+
         Schema::create('dismissers', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
@@ -17,6 +22,10 @@ return new class () extends Migration {
 
     public function down(): void
     {
+        if (App::environment() !== 'testing') {
+            return;
+        }
+
         Schema::dropIfExists('dismissers');
     }
 };
